refactor: adapt slash commands to new frontend architecture
Refactor slash commands implementation to follow the feature-based architecture:

- Move slash command types to features/assistant/types/
- Move slash command API client to features/assistant/api/
- Move CommandPalette component to features/assistant/components/
- Remove slash command exports from global API client and types
- Clean up debug logging from backend slash command service

The implementation now follows the modular architecture pattern where
slash commands are part of the assistant feature module rather than
being scattered across global files.
