+++ web/src/api/client.ts
@@ -4,8 +4,6 @@ import type {
   Session,
   CreateProjectRequest,
   CreateSessionRequest,
-  SlashCommand,
-  CommandResult,
 } from '../types'
 
 // Get auth credentials from localStorage or environment
@@ -119,24 +117,6 @@ export interface ChatMessage {
   tool_input?: string
   tool_use_id?: string
   tool_content?: string
-}
-
-// Slash command API methods
-export const slashCommands = {
-  // Get available slash commands for a session
-  getCommands: async (sessionId: number): Promise<SlashCommand[]> => {
-    const response = await api.get<SlashCommand[]>(`/sessions/${sessionId}/commands`)
-    return response.data
-  },
-
-  // Execute a slash command
-  executeCommand: async (sessionId: number, command: string, args?: string): Promise<CommandResult> => {
-    const response = await api.post<CommandResult>(`/sessions/${sessionId}/commands`, {
-      command,
-      args: args || '',
-    })
-    return response.data
-  },
 }
 
 export default api
