+++ web/src/features/assistant/components/CommandPalette.tsx
