+++ web/src/features/assistant/types/index.ts
@@ -1,0 +1,13 @@
+export interface SlashCommand {
+  name: string
+  description: string
+  path?: string
+  is_builtin: boolean
+  category?: string
+  arguments?: string[]
+}
+
+export interface CommandResult {
+  type: 'clear_chat' | 'show_modal' | 'claude_message' | 'error' | 'show_help' | 'status'
+  data: any
+}
