+++ web/src/types/index.ts
@@ -41,21 +41,6 @@ export interface Session {
   last_viewed_at?: string
 }
 
-export interface SlashCommand {
-  name: string
-  description: string
-  path?: string
-  is_builtin: boolean
-  category?: string
-  arguments?: string[]
-}
-
-export interface CommandResult {
-  type: 'clear_chat' | 'show_modal' | 'claude_message' | 'error' | 'show_help' | 'status'
-  data: any
-}
-
-
 export interface CreateProjectRequest {
   name: string
   path: string
